refactor(stella-graph): move parse's tests under the module they test
The last file #5420 left behind. `parse.rs` had an inline `#[cfg(test)] mod
tests`, so `mod tests;` could not also be declared and the suite beside it had
nowhere to go; the inline block moves out to `parse/tests.rs` and
`added_language_tests.rs` becomes a topic module under it.

Two things made this worth doing carefully rather than by eye. The inline block
sits in the *middle* of the file with production code after it, so the closing
brace is the one a scan finds at line 1152, not the last column-0 `}` in the
file — picking the latter splits a production function in half. And the block
holds a raw-string fixture that starts at column 0, so a uniform dedent shifts
its siblings and not it, silently rewriting a fixture for a crate that parses
whitespace-sensitive languages. The body is therefore copied verbatim and
reindented by rustfmt, which leaves string literal contents alone.

Checked rather than assumed: all 99 string literals in the block are
byte-identical either side of the move, and `cargo test -p stella-graph` reports
233 passed before and 233 after.

`parse.rs` drops from 1295 lines to 842.

Closes #5421
